Detail in clean.sh and adding nca/README

+++ nca/bin/clean.sh
@@ -3,5 +3,5 @@
 # This script only cleans all output and backups (BE CAREFUL)
 HERE=$(dirname "$0")
 
-rm -rf "$HERE"/backup_sig
-rm -rf "$HERE"/output
+[ -d "$HERE"/backup_sig ] && rm -rf "$HERE"/backup_sig
+[ -d "$HERE"/output ] && rm -rf "$HERE"/output
